+++ tests/e2e/specs/front-end/test-checkout-create-account.js
@@ -1,0 +1,6 @@
+/*
+ * Internal dependencies
+ */
+const { runCheckoutCreateAccountTest } = require( '@woocommerce/e2e-core-tests' );
+
+runCheckoutCreateAccountTest();