+++ tests/e2e/core-tests/CHANGELOG.md
@@ -28,6 +28,7 @@
 - Merchant Settings Shipping Zones
 - Shopper Variable product info updates on different variations
 - Merchant order emails flow
+- Shopper Checkout Create Account
 
 ## Fixed